comentários e pag not found

+++ controller/Tarefa.php
@@ -1,10 +1,10 @@
 <?php
-
 require_once('Controller.php');
 require_once('model' . DIRECTORY_SEPARATOR . 'TarefaModel.php');
 
 /**
-* 
+*@author Júlia Catarina R. e A. de Almeida <juc.amarante@gmail.com>
+* Classe Tarefa representando um Controller para controlar as regras de negócio
 */
 class Tarefa extends Controller
 {
@@ -15,6 +15,7 @@ class Tarefa extends Controller
 	public static function create()
 	{
 
+		//Redireciona para a view do formulário 
 		include('./view'. DIRECTORY_SEPARATOR . 'create.php');
 	}
 
@@ -55,7 +56,7 @@ class Tarefa extends Controller
 
 	/**
 	*@author Júlia Catarina R. e A. de Almeida <juc.amarante@gmail.com>
-	* Método 
+	* Método update responsável por chamar a view update apresentando os dados cadastrados
 	*/
 	public static function update($id)
 	{
@@ -68,9 +69,7 @@ class Tarefa extends Controller
 		{
 			self::iniciaSessao();
 		}
-
 		$_SESSION['dados'] = $dados;
-
 
 		//Retorna mensagem de sucesso
 		include('./view'. DIRECTORY_SEPARATOR . 'update.php');
@@ -78,7 +77,7 @@ class Tarefa extends Controller
 
 	/**
 	*@author Júlia Catarina R. e A. de Almeida <juc.amarante@gmail.com>
-	* Método do controller chamar o respectivo model
+	* Método atualiza responsavel por chamar o respectivo model para realizar a exclusão de tarefas
 	*/
 	public static function atualiza()
 	{
@@ -90,17 +89,28 @@ class Tarefa extends Controller
 		echo '<script> alert("Tarefa atualizada com sucesso");</script>';
 	}
 
-
+	/**
+	*@author Júlia Catarina R. e A. de Almeida <juc.amarante@gmail.com>
+	* Método delete responsavel por chamar o respectivo model para realizar a exclusão de tarefas
+	*/
 	public static function delete($id)
 	{
-		//Instanciando a classe responsável pelo model e passando os dados do formulários para a query
-		$model = new TarefaModel();
-		$dados = $model->delete("`id_tarefa` = " . $id);			
-		self::iniciaSessao();
+		//Verifica se existe um ID válido
+		if($_GET["id"] != null){
+			//Instanciando a classe responsável pelo model e passando os dados do formulários para a query
+			$model = new TarefaModel();
+			$dados = $model->delete("`id_tarefa` = " . $id);			
+			self::iniciaSessao();
 
-		//Retorna mensagem de sucesso
-		self::read();
-		echo '<script> alert("Tarefa excluida com sucesso");</script>';
+			//Retorna mensagem de sucesso
+			self::read();
+			echo '<script> alert("Tarefa excluida com sucesso");</script>';
+		}
+		//Se o ID não for válido manda para a página not found
+		else
+		{
+			include('./view'. DIRECTORY_SEPARATOR . 'notfound.php');
+		}
 	}
 }
